+++ src/components/features/cart/price-summary.tsx
@@ -4,7 +4,7 @@ import {
   getDeliveryFee,
   getTotalPrice,
 } from '@/lib/utils/common.util';
-import useCartStore from '@/store/cart/cart-store';
+import useCartStore from '@/store/cart-store';
 import { Equal, Plus } from 'lucide-react';
 
 export default function PriceSummary() {
