+++ src/components/features/detail/detail-button-area.tsx
@@ -3,7 +3,7 @@
 import { Button } from '@/components/ui/button';
 import LikeButton from './button-area/detail-like-button';
 import { ButtonAreaProps } from '@/types/detail.type';
-import useCartStore from '@/store/cart/cart-store';
+import useCartStore from '@/store/cart-store';
 
 export default function ButtonArea({
   id,
