+++ src/components/features/order/order-info.tsx
@@ -3,7 +3,7 @@ import {
   getDeliveryFee,
   getTotalPrice,
 } from '@/lib/utils/common.util';
-import useCartStore from '@/store/cart/cart-store';
+import useCartStore from '@/store/cart-store';
 import { Plus } from 'lucide-react';
 
 const styles = {
